fix: 修复 save_predictions 中 float32 类型 JSON 序列化失败
numpy float32 无法被 json.dump 序列化，显式转换为 Python float。

+++ scripts/predict_stocks.py
@@ -542,7 +542,7 @@ def save_predictions(all_forward):
     snapshot = {}
     for code, info in all_forward.items():
         if info.get("rows"):
-            snapshot[code] = round(info["rows"][-1]["cum_chg"], 4)
+            snapshot[code] = float(round(info["rows"][-1]["cum_chg"], 4))
     try:
         os.makedirs(os.path.dirname(STABILITY_CACHE), exist_ok=True)
         tmp_path = STABILITY_CACHE + ".tmp"
